+++ QoL/GlobeTweaks/GlyphModule.cs
@@ -1,5 +1,4 @@
-﻿using PhoenixPoint.Geoscape.Entities;
-using PhoenixPoint.Geoscape.View;
+﻿using PhoenixPoint.Geoscape.View;
 using System;
 using UnityEngine;
 
@@ -9,15 +8,26 @@ namespace Sheepy.PhoenixPt.GlobeTweaks {
       private static IPatch FovPatch;
 
       public void DoPatches () {
-         FovPatch = TryPatch( typeof( GeoSiteVisualsController ), "Awake", postfix: nameof( Test2 ) );
+         var conf = Mod.Config.Haven_Icons ?? new HavenIconConfig();
+         if ( conf.Always_Show_Recruit || conf.Always_Show_Soldier || conf.Always_Show_Trade )
+            FovPatch = TryPatch( typeof( GeoSiteVisualsController ), "Awake", postfix: nameof( GeoSiteVisualsController_AfterAwake ) );
       }
 
-      private static void Test2 ( GeoSiteVisualsController __instance ) { try {
-         var fov = __instance.RecruitAvailableIcon.transform.parent.gameObject.GetComponent<FovControllableBehavior>();
-         if ( fov == null || fov.ControllingDef.InvisibleOverFov < 0 ) return;
-         fov.ControllingDef.InvisibleOverFov = -1;
-         Api( "zy.ui.dump", __instance.RecruitAvailableIcon.transform.parent.parent );
+      private static void GeoSiteVisualsController_AfterAwake ( GeoSiteVisualsController __instance ) { try {
+         var conf = Mod.Config.Haven_Icons ?? new HavenIconConfig();
+         if ( conf.Always_Show_Recruit ) DisableFov( __instance.RecruitAvailableIcon .transform.parent, "RecruitAvailableIcon"  );
+         if ( conf.Always_Show_Soldier ) DisableFov( __instance.SoldiersAvailableIcon.transform.parent, "SoldiersAvailableIcon" );
+         if ( conf.Always_Show_Trade   ) DisableFov( __instance.SuppliesResourcesIcon.transform.parent, "SuppliesResourcesIcon" );
+         //Api( "zy.ui.dump", __instance.RecruitAvailableIcon.transform.parent.parent ); // Dump gui tree with debug console
          FovPatch.Unpatch();
       } catch ( Exception ex ) { Error( ex ); } }
+
+      private static void DisableFov ( Transform t, string type ) {
+         var fov = t.gameObject.GetComponent<FovControllableBehavior>();
+         if ( fov?.ControllingDef?.InvisibleOverFov > 0 )
+            fov.ControllingDef.InvisibleOverFov = -1;
+         else
+            Warn( "Fov not found, cannot set always visible: {0}", type );
+      }
    }
 }